Fixed automatic login script editing upon installation.
Previously, installation of a new version of MESA would attempt to adjust the MESA_DIR environment variable declaration in a login script. It would also add in a comment explanining the change, but it inserted it in the wrong place and didn't add a newline character, causing it to comment out lines in the login script. This has been fixed.