Catch another MesaScript Error
Catch TypeErrors caused by loading MesaScript without a MESA installation.

Previously, if mesascript was installed, but MESA was not, a TypeError caused by being unable to determine the version number would halt any use of the CLI whatsoever. Now users should be able to access it to get help on available commands and use the install method.